refactor : extract initial admin creation into shared helper
